pyproject.toml
[tool.poetry]
name = "bot"
version = "1.0.0"
description = "A Telegram bot that provides statistics about files."
authors = ["Francesco146"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.13.1"
python-telegram-bot = "21.9"
humanize = "4.11.0"
python-dotenv = "1.0.1"
argparse = "1.4.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"